chore: cleanup app route issues and added zenboard as a protected route
