+++ src/app/router/protected.ts
@@ -1,0 +1,6 @@
+import { Router } from "express";
+import { zenboardRouter } from "../../zenboard";
+
+export const protectedRouter = Router();
+
+protectedRouter.use("/zenboard", zenboardRouter);
