+++ src/zenboard/models/index.ts
@@ -1,0 +1,1 @@
+// each model will be separated in files and exported here
